+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,0 +1,108 @@
+package kitchenpos.order.application;
+
+import kitchenpos.menu.dao.MenuDao;
+import kitchenpos.order.dao.OrderDao;
+import kitchenpos.order.dao.OrderLineItemDao;
+import kitchenpos.order.dao.OrderTableDao;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
+@Service
+public class OrderService {
+    private final MenuDao menuDao;
+    private final OrderDao orderDao;
+    private final OrderLineItemDao orderLineItemDao;
+    private final OrderTableDao orderTableDao;
+
+    public OrderService(
+            final MenuDao menuDao,
+            final OrderDao orderDao,
+            final OrderLineItemDao orderLineItemDao,
+            final OrderTableDao orderTableDao
+    ) {
+        this.menuDao = menuDao;
+        this.orderDao = orderDao;
+        this.orderLineItemDao = orderLineItemDao;
+        this.orderTableDao = orderTableDao;
+    }
+
+    @Transactional
+    public Order create(final Order order) {
+        final List<OrderLineItem> orderLineItems = order.getOrderLineItems();
+
+        if (CollectionUtils.isEmpty(orderLineItems)) {
+            throw new IllegalArgumentException();
+        }
+
+        final List<Long> menuIds = orderLineItems.stream()
+                .map(OrderLineItem::getMenuId)
+                .collect(Collectors.toList());
+
+        if (orderLineItems.size() != menuDao.countByIdIn(menuIds)) {
+            throw new IllegalArgumentException();
+        }
+
+        final OrderTable orderTable = orderTableDao.findById(order.getOrderTableId())
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (orderTable.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+
+        order.setOrderTableId(orderTable.getId());
+        order.setOrderStatus(OrderStatus.COOKING.name());
+        order.setOrderedTime(LocalDateTime.now());
+
+        final Order savedOrder = orderDao.save(order);
+
+        final Long orderId = savedOrder.getId();
+        final List<OrderLineItem> savedOrderLineItems = new ArrayList<>();
+        for (final OrderLineItem orderLineItem : orderLineItems) {
+            orderLineItem.setOrderId(orderId);
+            savedOrderLineItems.add(orderLineItemDao.save(orderLineItem));
+        }
+        savedOrder.setOrderLineItems(savedOrderLineItems);
+
+        return savedOrder;
+    }
+
+    public List<Order> list() {
+        final List<Order> orders = orderDao.findAll();
+
+        for (final Order order : orders) {
+            order.setOrderLineItems(orderLineItemDao.findAllByOrderId(order.getId()));
+        }
+
+        return orders;
+    }
+
+    @Transactional
+    public Order changeOrderStatus(final Long orderId, final Order order) {
+        final Order savedOrder = orderDao.findById(orderId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (Objects.equals(OrderStatus.COMPLETION.name(), savedOrder.getOrderStatus())) {
+            throw new IllegalArgumentException();
+        }
+
+        final OrderStatus orderStatus = OrderStatus.valueOf(order.getOrderStatus());
+        savedOrder.setOrderStatus(orderStatus.name());
+
+        orderDao.save(savedOrder);
+
+        savedOrder.setOrderLineItems(orderLineItemDao.findAllByOrderId(orderId));
+
+        return savedOrder;
+    }
+}
