+++ src/main/java/kitchenpos/order/application/TableService.java
@@ -1,0 +1,73 @@
+package kitchenpos.order.application;
+
+import kitchenpos.order.dao.OrderDao;
+import kitchenpos.order.dao.OrderTableDao;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+
+@Service
+public class TableService {
+    private final OrderDao orderDao;
+    private final OrderTableDao orderTableDao;
+
+    public TableService(final OrderDao orderDao, final OrderTableDao orderTableDao) {
+        this.orderDao = orderDao;
+        this.orderTableDao = orderTableDao;
+    }
+
+    @Transactional
+    public OrderTable create(final OrderTable orderTable) {
+        orderTable.setTableGroupId(null);
+
+        return orderTableDao.save(orderTable);
+    }
+
+    public List<OrderTable> list() {
+        return orderTableDao.findAll();
+    }
+
+    @Transactional
+    public OrderTable changeEmpty(final Long orderTableId, final OrderTable orderTable) {
+        final OrderTable savedOrderTable = orderTableDao.findById(orderTableId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (Objects.nonNull(savedOrderTable.getTableGroupId())) {
+            throw new IllegalArgumentException();
+        }
+
+        if (orderDao.existsByOrderTableIdAndOrderStatusIn(
+                orderTableId, Arrays.asList(OrderStatus.COOKING.name(), OrderStatus.MEAL.name()))) {
+            throw new IllegalArgumentException();
+        }
+
+        savedOrderTable.setEmpty(orderTable.isEmpty());
+
+        return orderTableDao.save(savedOrderTable);
+    }
+
+    @Transactional
+    public OrderTable changeNumberOfGuests(final Long orderTableId, final OrderTable orderTable) {
+        final int numberOfGuests = orderTable.getNumberOfGuests();
+
+        if (numberOfGuests < 0) {
+            throw new IllegalArgumentException();
+        }
+
+        final OrderTable savedOrderTable = orderTableDao.findById(orderTableId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (savedOrderTable.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+
+        savedOrderTable.setNumberOfGuests(numberOfGuests);
+
+        return orderTableDao.save(savedOrderTable);
+    }
+}
