+++ src/main/java/kitchenpos/order/dao/OrderDao.java
@@ -1,0 +1,18 @@
+package kitchenpos.order.dao;
+
+import kitchenpos.order.domain.Order;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface OrderDao {
+    Order save(Order entity);
+
+    Optional<Order> findById(Long id);
+
+    List<Order> findAll();
+
+    boolean existsByOrderTableIdAndOrderStatusIn(Long orderTableId, List<String> orderStatuses);
+
+    boolean existsByOrderTableIdInAndOrderStatusIn(List<Long> orderTableIds, List<String> orderStatuses);
+}
