+++ src/main/java/kitchenpos/order/dao/OrderLineItemDao.java
@@ -1,0 +1,16 @@
+package kitchenpos.order.dao;
+
+import kitchenpos.order.domain.OrderLineItem;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface OrderLineItemDao {
+    OrderLineItem save(OrderLineItem entity);
+
+    Optional<OrderLineItem> findById(Long id);
+
+    List<OrderLineItem> findAll();
+
+    List<OrderLineItem> findAllByOrderId(Long orderId);
+}
