+++ src/main/java/kitchenpos/order/dao/OrderTableDao.java
@@ -1,0 +1,18 @@
+package kitchenpos.order.dao;
+
+import kitchenpos.order.domain.OrderTable;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface OrderTableDao {
+    OrderTable save(OrderTable entity);
+
+    Optional<OrderTable> findById(Long id);
+
+    List<OrderTable> findAll();
+
+    List<OrderTable> findAllByIdIn(List<Long> ids);
+
+    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
+}
