+++ src/main/java/kitchenpos/order/dao/TableGroupDao.java
@@ -1,0 +1,14 @@
+package kitchenpos.order.dao;
+
+import kitchenpos.order.domain.TableGroup;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface TableGroupDao {
+    TableGroup save(TableGroup entity);
+
+    Optional<TableGroup> findById(Long id);
+
+    List<TableGroup> findAll();
+}
