+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,0 +1,52 @@
+package kitchenpos.order.domain;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public class Order {
+    private Long id;
+    private Long orderTableId;
+    private String orderStatus;
+    private LocalDateTime orderedTime;
+    private List<OrderLineItem> orderLineItems;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(final Long id) {
+        this.id = id;
+    }
+
+    public Long getOrderTableId() {
+        return orderTableId;
+    }
+
+    public void setOrderTableId(final Long orderTableId) {
+        this.orderTableId = orderTableId;
+    }
+
+    public String getOrderStatus() {
+        return orderStatus;
+    }
+
+    public void setOrderStatus(final String orderStatus) {
+        this.orderStatus = orderStatus;
+    }
+
+    public LocalDateTime getOrderedTime() {
+        return orderedTime;
+    }
+
+    public void setOrderedTime(final LocalDateTime orderedTime) {
+        this.orderedTime = orderedTime;
+    }
+
+    public List<OrderLineItem> getOrderLineItems() {
+        return orderLineItems;
+    }
+
+    public void setOrderLineItems(final List<OrderLineItem> orderLineItems) {
+        this.orderLineItems = orderLineItems;
+    }
+}
