+++ src/main/java/kitchenpos/order/domain/OrderTable.java
@@ -1,0 +1,40 @@
+package kitchenpos.order.domain;
+
+public class OrderTable {
+    private Long id;
+    private Long tableGroupId;
+    private int numberOfGuests;
+    private boolean empty;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(final Long id) {
+        this.id = id;
+    }
+
+    public Long getTableGroupId() {
+        return tableGroupId;
+    }
+
+    public void setTableGroupId(final Long tableGroupId) {
+        this.tableGroupId = tableGroupId;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public void setNumberOfGuests(final int numberOfGuests) {
+        this.numberOfGuests = numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+
+    public void setEmpty(final boolean empty) {
+        this.empty = empty;
+    }
+}
