+++ src/main/java/kitchenpos/order/domain/TableGroup.java
@@ -1,0 +1,34 @@
+package kitchenpos.order.domain;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public class TableGroup {
+    private Long id;
+    private LocalDateTime createdDate;
+    private List<OrderTable> orderTables;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(final Long id) {
+        this.id = id;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
+    }
+
+    public void setCreatedDate(final LocalDateTime createdDate) {
+        this.createdDate = createdDate;
+    }
+
+    public List<OrderTable> getOrderTables() {
+        return orderTables;
+    }
+
+    public void setOrderTables(final List<OrderTable> orderTables) {
+        this.orderTables = orderTables;
+    }
+}
