+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -1,0 +1,42 @@
+package kitchenpos.order.ui;
+
+import kitchenpos.order.application.OrderService;
+import kitchenpos.order.domain.Order;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class OrderRestController {
+    private final OrderService orderService;
+
+    public OrderRestController(final OrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    @PostMapping("/api/orders")
+    public ResponseEntity<Order> create(@RequestBody final Order order) {
+        final Order created = orderService.create(order);
+        final URI uri = URI.create("/api/orders/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @GetMapping("/api/orders")
+    public ResponseEntity<List<Order>> list() {
+        return ResponseEntity.ok()
+                .body(orderService.list())
+                ;
+    }
+
+    @PutMapping("/api/orders/{orderId}/order-status")
+    public ResponseEntity<Order> changeOrderStatus(
+            @PathVariable final Long orderId,
+            @RequestBody final Order order
+    ) {
+        return ResponseEntity.ok(orderService.changeOrderStatus(orderId, order));
+    }
+}
