+++ src/main/java/kitchenpos/order/ui/TableGroupRestController.java
@@ -1,0 +1,34 @@
+package kitchenpos.order.ui;
+
+import kitchenpos.order.application.TableGroupService;
+import kitchenpos.order.domain.TableGroup;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+
+@RestController
+public class TableGroupRestController {
+    private final TableGroupService tableGroupService;
+
+    public TableGroupRestController(final TableGroupService tableGroupService) {
+        this.tableGroupService = tableGroupService;
+    }
+
+    @PostMapping("/api/table-groups")
+    public ResponseEntity<TableGroup> create(@RequestBody final TableGroup tableGroup) {
+        final TableGroup created = tableGroupService.create(tableGroup);
+        final URI uri = URI.create("/api/table-groups/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @DeleteMapping("/api/table-groups/{tableGroupId}")
+    public ResponseEntity<Void> ungroup(@PathVariable final Long tableGroupId) {
+        tableGroupService.ungroup(tableGroupId);
+        return ResponseEntity.noContent()
+                .build()
+                ;
+    }
+}
