+++ src/test/java/kitchenpos/menu/ui/MenuGroupRestControllerTest.java
@@ -3,8 +3,7 @@ package kitchenpos.menu.ui;
 import kitchenpos.menu.application.MenuGroupService;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.common.fixtrue.MenuGroupFixture;
-import kitchenpos.menu.ui.MenuGroupRestController;
-import kitchenpos.ui.RestControllerTest;
+import kitchenpos.common.ui.RestControllerTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
