+++ src/test/java/kitchenpos/menu/ui/MenuRestControllerTest.java
@@ -8,7 +8,7 @@ import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
-import kitchenpos.ui.RestControllerTest;
+import kitchenpos.common.ui.RestControllerTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
