+++ src/test/java/kitchenpos/product/ui/ProductRestControllerTest.java
@@ -3,7 +3,7 @@ package kitchenpos.product.ui;
 import kitchenpos.common.fixtrue.ProductFixture;
 import kitchenpos.product.application.ProductService;
 import kitchenpos.product.domain.Product;
-import kitchenpos.ui.RestControllerTest;
+import kitchenpos.common.ui.RestControllerTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
